# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Receipt Assistant API",
    "version": "2.0.0-alpha",
    "description": "Double-entry ledger backend. The v1 resources (transactions, postings, accounts, documents) are under active construction — see issues #33 (schema), #35 (transactions API), #36 (accounts API)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Violation": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "code"
        ]
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          },
          "violations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Violation"
            }
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      },
...
      "Document": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
          },
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
          },
          "kind": {
            "type": "string",
            "enum": [
              "receipt_image",
              "receipt_email",
              "receipt_pdf",
              "statement_pdf",
              "other"
            ]
          },
          "file_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "mime_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "sha256": {
            "type": "string"
          },
          "ocr_text": {
            "type": [
              "string",
              "null"
            ]
          },
          "extraction_meta": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {}
          },
          "source_ingest_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
          },
          "deleted_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "kind",
          "file_path",
          "mime_type",
          "sha256",
          "ocr_text",
          "extraction_meta",
          "source_ingest_id",
          "deleted_at",
          "created_at",
          "updated_at"
        ]
      },
...
      "UploadDocumentForm": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary"
          },
          "kind": {
            "type": "string",
            "enum": [
              "receipt_image",
              "receipt_email",
              "receipt_pdf",
              "statement_pdf",
              "other"
            ]
          }
        }
      },
      "CreateDocumentLinkRequest": {
        "type": "object",
        "properties": {
          "transaction_id": {
            "type": "string",
            "format": "uuid",
            "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
          }
        },
        "required": [
          "transaction_id"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/documents": {
      "post": {
        "summary": "Upload a document (multipart, sha256 content-dedup)",
        "tags": [
          "documents"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadDocumentForm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Duplicate upload detected — returns the existing document row",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "201": {
            "description": "Document created",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{id}": {
      "get": {
        "summary": "Get document metadata",
        "tags": [
          "documents"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false",
                "1",
                "0"
              ],
              "description": "Include soft-deleted documents in the response. Default false."
            },
            "required": false,
            "name": "include_deleted",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Document",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "304": {
            "description": "Not modified (If-None-Match matched)"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a document. Soft-deletes by default; ?hard=true removes the row + file; ?cascade=true also handles linked transactions.",
        "description": "Default: soft delete (sets deleted_at). ?hard=true: hard delete (row + file); requires no remaining links unless ?cascade=true is also set. ?cascade=true: linked posted transactions are voided, draft/error transactions are hard-deleted, voided transactions are left intact, reconciled transactions abort the operation with 409. ?cascade=true&hard=true: every linked transaction is hard-deleted (postings cascade), the document is hard-deleted, and the image file is removed. Reconciled transactions always block hard cascades — unreconcile first.",
        "tags": [
          "documents"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false",
                "1",
                "0"
              ],
              "description": "Hard-delete the row and the on-disk image. Default false."
            },
            "required": false,
            "name": "hard",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false",
                "1",
                "0"
              ],
              "description": "Also process transactions linked to this document. Combine with hard=true for full purge."
            },
            "required": false,
            "name": "cascade",
            "in": "query"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Hard delete refused because document has links (use cascade), or cascade refused because a linked transaction is reconciled.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{id}/content": {
      "get": {
        "summary": "Stream document binary content",
        "tags": [
          "documents"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Raw file bytes",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{id}/links": {
      "post": {
        "summary": "Link a document to a transaction (idempotent)",
        "tags": [
          "documents"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentLinkRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Linked (or already linked — idempotent)"
          },
          "404": {
            "description": "Document or transaction not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{id}/links/{txn_id}": {
      "delete": {
        "summary": "Unlink a document from a transaction",
        "tags": [
          "documents"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": true,
            "name": "txn_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Unlinked"
          },
          "404": {
            "description": "Link not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{id}/restore": {
      "post": {
        "summary": "Restore a soft-deleted document",
        "tags": [
          "documents"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Restored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "404": {
            "description": "Document not found or not deleted",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}